==> scripts/2_Fishing_displacement.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(here)
library(tidyverse)

yoi<-c(2007,2008,2009,2010,2011)
```


# Load data

Fishing data

```{r}
fish<-readRDS(here("output_data/Fishing_data_clean.Rdata"))
```

Management data

```{r}
me<-readRDS(here("output_data/Management_data_clean.Rdata")) 
```

Join data

```{r}
fish_me<-full_join(fish, me, by = c("Fished_ISO3" = "Export_ISO3"))   %>% 
  mutate(Fishing_ISO3 = as.character(Fishing_ISO3),
         Fished_ISO3 = as.character(Fished_ISO3)) %>% 
  filter(Year>=1976,
         Year <= 2015) 

test<-fish_me %>% 
  filter(is.na(ME_score) == T)

# These countries do not have an ME score
unique(test$Fished_ISO3)
```

```{r}
aus<-fish_me %>% 
  filter(Fishing_ISO3 == "AUS")
```

# Total fishing

```{r}
fishing<-fish_me %>% 
  group_by(Fishing_ISO3, Year) %>% 
  summarise(Total_fished = sum(Total_fished, na.rm = T))

fishing_iso3<-fish_me %>% 
  group_by(Fishing_ISO3, Fished_ISO3, Year) %>% 
  summarise(Total_fished_where = sum(Total_fished, na.rm = T))
```

# Displaced (fished) by country

EEZ and High Seas

```{r}
fish_displaced<- fish_me %>% 
  filter(!Fishing_ISO3 == Fished_ISO3) %>% 
  group_by(Fishing_ISO3, Fished_ISO3, Year) %>% 
  summarise(disp_fished = sum(Total_fished, na.rm = T)) 

all_fish_displaced<-fish_displaced %>% 
  group_by(Fishing_ISO3, Year) %>% 
  summarise(disp_fished = sum(disp_fished, na.rm = T))


test<-all_fish_displaced %>% 
  filter(Fishing_ISO3 == "AGO",
         Year %in% yoi)
```



# Unequal Fishing Displacement

Displacement (fishing) from countries with higher management fishing in lower management waters

```{r}
unequal<-fish_me %>% 
  filter(!Fishing_ISO3 == Fished_ISO3) %>% 
  full_join(., me, by = c("Fishing_ISO3" = "Export_ISO3")) %>% 
  rename(Fished_ME = ME_score.x,
         Fishing_ME = ME_score.y) 

sub_unequal<-unequal %>% 
  filter(Fished_ME < Fishing_ME)

unequal_displaced<- sub_unequal %>% 
  filter(!Fishing_ISO3 == Fished_ISO3) %>% 
  group_by(Fishing_ISO3, Fished_ISO3, Year) %>% 
  summarise(unequal_disp_fished = sum(Total_fished, na.rm = T)) 

unequal_fish<- sub_unequal %>% 
  group_by(Fishing_ISO3, Year) %>% 
  summarise(unequal_fish = sum(Total_fished, na.rm = T))
  
```

High Seas only

```{r}
unequal_hs<-fish_me %>% 
  filter(!Fishing_ISO3 == Fished_ISO3,
         Fished_ISO3 == "HighSeas") %>% 
  full_join(., me, by = c("Fishing_ISO3" = "Export_ISO3")) %>% 
  rename(Fished_ME = ME_score.x,
         Fishing_ME = ME_score.y) 

sub_unequal_hs<-unequal_hs %>% 
  filter(Fished_ME < Fishing_ME)

unequal_displaced_hs<- sub_unequal_hs %>% 
  filter(!Fishing_ISO3 == Fished_ISO3) %>% 
  group_by(Fishing_ISO3, Fished_ISO3, Year) %>% 
  summarise(unequal_disp_fished_hs = sum(Total_fished, na.rm = T)) 

unequal_fish_hs<- sub_unequal_hs %>% 
  group_by(Fishing_ISO3, Year) %>% 
  summarise(unequal_fish_hs = sum(Total_fished, na.rm = T))
  
```

# Reverse displacement

Displacement (fishing) from countries with lower management fishing in higher management waters

Intuitively, reverse displacement should equal displacement - unequal displacement. This is mostly the case, except when some countries do not have ME scores then that catch gets dropped. See BEL in 1976 as example.

```{r}
rev_unequal<-unequal %>% 
  filter(Fished_ME >= Fishing_ME)

rev_displaced<- rev_unequal %>% 
  filter(!Fishing_ISO3 == Fished_ISO3) %>% 
  group_by(Fishing_ISO3, Fished_ISO3, Year) %>% 
  summarise(rev_disp_fished = sum(Total_fished, na.rm = T)) 

rev_fish<- rev_unequal %>% 
  group_by(Fishing_ISO3, Year) %>% 
  summarise(rev_fish = sum(Total_fished, na.rm = T))
  
```

High Seas only - this doesn't occur because high seas doesn't fish

```{r}
rev_hs<-fish_me %>% 
  filter(!Fishing_ISO3 == Fished_ISO3,
         Fished_ISO3 == "HighSeas") %>% 
  full_join(., me, by = c("Fishing_ISO3" = "Export_ISO3")) %>% 
  rename(Fished_ME = ME_score.x,
         Fishing_ME = ME_score.y) 

rev_unequal_hs<-rev_hs %>% 
  filter(Fished_ME >= Fishing_ME)

rev_displaced_hs<- rev_unequal_hs %>% 
  filter(!Fishing_ISO3 == Fished_ISO3) %>% 
  group_by(Fishing_ISO3, Fished_ISO3, Year) %>% 
  summarise(unequal_disp_fished_hs = sum(Total_fished, na.rm = T)) 

rev_fish_hs<- rev_unequal_hs %>% 
  group_by(Fishing_ISO3, Year) %>% 
  summarise(unequal_fish_hs = sum(Total_fished, na.rm = T))
  
```

# Unequal fishing in each countries waters

```{r}

unequal_waters<-sub_unequal %>% 
  group_by(Fished_ISO3, Year) %>% 
  summarise(unequal_fishing_in_waters = sum(Total_fished, na.rm = T))

unequal_waters_disp<-sub_unequal %>% 
  group_by(Fished_ISO3, Fishing_ISO3, Year) %>% 
  summarise(unequal_fishing_in_waters = sum(Total_fished, na.rm = T))
```

# Reverse fishing in each countries waters

```{r}
rev_waters<-rev_unequal %>% 
  group_by(Fished_ISO3, Year) %>% 
  summarise(rev_fishing_in_waters = sum(Total_fished, na.rm = T))

rev_waters_disp<-rev_unequal %>% 
  group_by(Fished_ISO3, Fishing_ISO3, Year) %>% 
  summarise(rev_fishing_in_waters = sum(Total_fished, na.rm = T))
```

# Final table
Join 

```{r}
fish_results<-full_join(fishing, all_fish_displaced) %>% 
  full_join(., unequal_fish) %>% 
  full_join(., unequal_fish_hs) %>% 
  full_join(., unequal_waters, by = c("Fishing_ISO3" = "Fished_ISO3", "Year")) %>% 
  full_join(., rev_fish) %>% 
  full_join(., rev_waters, by = c("Fishing_ISO3" = "Fished_ISO3", "Year")) %>% 
  mutate(unequal_fish_hs =ifelse(is.na(unequal_fish_hs) == T, 0, as.numeric(unequal_fish_hs)),
        unequal_fish_eez = unequal_fish - unequal_fish_hs,
        prop_disp = disp_fished/Total_fished,
        prop_unequal_disp = unequal_fish/disp_fished,
        prop_rev_disp = rev_fish/disp_fished) 
```

# Table with country partners


```{r}
fish_results_country<-full_join(fishing_iso3, fish_displaced) %>% 
  full_join(., unequal_displaced) %>% 
  full_join(., unequal_displaced_hs) %>% 
  full_join(., unequal_waters_disp, by = c("Fishing_ISO3" = "Fished_ISO3", "Fished_ISO3" = "Fishing_ISO3", "Year")) %>% 
  full_join(., rev_displaced) %>% 
  full_join(., rev_waters_disp, by = c("Fishing_ISO3" = "Fished_ISO3", "Fished_ISO3" = "Fishing_ISO3", "Year")) %>% 
  full_join(., me, by = c("Fishing_ISO3" = "Export_ISO3")) %>% 
  rename(Fishing_ME = ME_score,
         Fishing_quantile = quantile) %>% 
  full_join(., me, by = c("Fished_ISO3" = "Export_ISO3")) %>% 
  rename(Fished_ME = ME_score,
         Fished_quantile = quantile) 
```



```{r}
write_csv(fish_results, here("output_data/Fish_displacement.csv"))

write_csv(fish_results_country, here("output_data/Fish_displacement_by_country.csv"))
```

```{r}
test<-fish_results %>% 
  filter(!is.na(Total_fished) == T)

length(unique(test$Fishing_ISO3))

test<-test %>% 
  filter(Year %in% c(2007,2008,2009,2010,2011))


length(unique(test$Fishing_ISO3))

test2<-test %>% 
  filter(!is.na(disp_fished) == T)

length(unique(test2$Fishing_ISO3))

test2<-test2 %>% 
  filter(Year %in% c(2007,2008,2009,2010,2011))

length(unique(test2$Fishing_ISO3))

test3<-fish_results %>% 
  full_join(., me, by = c("Fishing_ISO3" = "Export_ISO3"))

test4<-test3 %>% 
  filter(!is.na(Total_fished) == T,
         !is.na(ME_score) == T)

length(unique(test4$Fishing_ISO3))

test4<-test4 %>% 
  filter(Year %in% c(2007,2008,2009,2010,2011))

length(unique(test4$Fishing_ISO3))
```

```{r}
test5<-read_csv(here("output_data/Trade_displacement.csv")) %>% 
  filter(!is.na(imports) == T,
         !is.na(Import_ISO3) == T,
         !is.na(ME_score) == T)

test6<-full_join(test5, test4, by = c("Import_ISO3" = "Fishing_ISO3","Year", "ME_score", "quantile"))

test7<-test6 %>% 
  filter(!is.na(imports) == T,
         Year %in% c(2007,2008,2009,2010,2011)) 

test8<-test6%>% 
  filter(!is.na(imports) == T,
         !is.na(ME_score) == T,
         Year %in% c(2007,2008,2009,2010,2011))

length(unique(test8$Import_ISO3))
         

length(unique(test7$Import_ISO3))
```

```{r}
test6<-test6 %>% 
  filter(!is.na(imports) == T,
         !is.na(Total_fished) == T,
         !is.na(ME_score) == T)

length(unique(test6$Import_ISO3))

test6<-test6 %>% 
  filter(!is.na(imports) == T,
         !is.na(Total_fished) == T,
         !is.na(ME_score) == T,
         Year %in% c(2007,2008,2009,2010,2011))

length(unique(test6$Import_ISO3))
```
