==> scripts/3_Trade_displacement.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Load libraries
```{r}
library(here)
library(tidyverse)

yoi<-c(2007,2008,2009,2010,2011)
```


# Load data

Trade data
```{r}
trade<-readRDS(here("output_data/Trade_data_clean.Rdata"))
```

Management data

```{r}
me<-readRDS(here("output_data/Management_data_clean.Rdata")) 
```

Join data

```{r}
trade_me<-full_join(trade, me, by = "Export_ISO3")

test<-trade_me %>% 
  filter(is.na(ME_score) == T)

# These countries do not have an ME score
unique(test$Export_ISO3)
```

# Displaced (imported) by country

```{r}
imports<- trade_me %>% 
  group_by(Import_ISO3, Year) %>% 
  summarise(imports = sum(Total_tonnage, na.rm = T))

imports_iso3<- trade_me %>% 
  group_by(Import_ISO3, Export_ISO3, Year) %>% 
  summarise(imports = sum(Total_tonnage, na.rm = T))
```

# Unequal Import Displacement

Displacement (imports) from countries with lower management

```{r}
unequal<-trade_me %>% 
  full_join(., me, by = c("Import_ISO3" = "Export_ISO3")) %>% 
  rename(Export_ME = ME_score.x,
         Import_ME = ME_score.y)

sub_unequal<-unequal %>% 
  filter(Export_ME < Import_ME)

unequal_imports<- sub_unequal %>% 
  group_by(Import_ISO3, Year) %>% 
  summarise(unequal_imports = sum(Total_tonnage, na.rm = T))

unequal_imports_iso3<- sub_unequal %>% 
  group_by(Import_ISO3, Export_ISO3, Year) %>% 
  summarise(unequal_imports = sum(Total_tonnage, na.rm = T))
  
```

Join 

```{r}
trade_results<-full_join(imports, unequal_imports) %>% 
  mutate(prop_unequal = unequal_imports/imports) %>% 
  full_join(., me, by = c("Import_ISO3" = "Export_ISO3"))

trade_results_iso3<-full_join(imports_iso3, unequal_imports_iso3) %>% 
  full_join(., me) %>% 
  rename(ME_score_exporter = ME_score) %>% 
  full_join(., me, by = c("Import_ISO3" = "Export_ISO3")) %>% 
  rename(ME_score_importer = ME_score) 
```

```{r}
write.csv(trade_results, here("output_data/Trade_displacement.csv"))

write.csv(trade_results_iso3, here("output_data/Trade_displacement_by_country.csv"))
```

```{r}
test<-read_csv(here("output_data/Trade_displacement.csv")) %>% 
  filter(!is.na(imports) == T,
         !is.na(Import_ISO3) == T)

length(unique(test$Import_ISO3))

test_me<-test %>% 
  filter(!is.na(ME_score) == T)

length(unique(test_me$Import_ISO3))

test2<-read_csv(here("output_data/Trade_displacement_by_country.csv")) %>% 
    filter(!is.na(imports) == T,
         !is.na(Import_ISO3) == T)

length(unique(test2$Import_ISO3))

test3<-test2 %>% 
  filter(!is.na(ME_score_importer) == T)

length(unique(test3$Import_ISO3))

test<-trade_results %>% 
  filter(Year %in% yoi,
         Import_ISO3 == "ALB")
```

```{r}
#test2<-imex_dat %>% 
#  filter(Year %in% c(2011,2012,2013,2014,2015),
#         Import_ISO3 == "ALB",
#         Source %in% c("Domestic Catch", "Foreign Fishing", "ReExport Domestic Catch", "ReExport Foreign #Fishing")) 
  
```
