==> scripts/8_Supplementary_Table1.Rmd <==
---
title: "Supplementary_Table1"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(dplyr)
library(countrycode)
library(reshape2)
library(stringr)
library(data.table)
library(zoo)
library(here)

yoi<-c(2007,2008,2009,2010,2011)
```

# Data

```{r}
trade<-read.csv(here("output_data/Trade_displacement.csv"))

fish<-read.csv(here("output_data/Fish_displacement.csv")) 

me_dat<-readRDS(here("output_data/Management_data_clean.Rdata")) 

fish_fmi_wadj<-read.csv(here("output_data/Fish_displacement_FMI_wadj.csv")) 

fish_fmi<-read.csv(here("output_data/Fish_displacement_FMI.csv")) 

fmi_dat<-readRDS(here("output_data/FMI_clean.Rdata")) %>% 
  dplyr::select(Export_ISO3, FMI_weight_adj, FMI)
```

# Subset trade and fisheries data to average between 2007-2011

```{r}
sub_trade<-trade %>% 
  filter(Year %in% yoi) %>% 
  dplyr::select(-prop_unequal) %>% 
  group_by(Import_ISO3) %>% 
  summarise_all(funs(mean), na.rm = TRUE) %>% 
  dplyr::select(-Year) 

```

```{r}
sub_fish<-fish %>% 
  filter(Year %in% yoi) %>% 
  dplyr::select(-prop_disp, -prop_unequal_disp, -prop_rev_disp) %>% 
  group_by(Fishing_ISO3) %>% 
  summarise_all(funs(mean), na.rm = TRUE) %>% 
  dplyr::select(-Year) 
```

```{r}
sub_fish_fmi_wadj<-fish_fmi_wadj %>% 
  filter(Year %in% yoi) %>% 
  dplyr::select(-prop_disp, -prop_unequal_disp, -prop_rev_disp) %>% 
  group_by(Fishing_ISO3) %>% 
  summarise_all(funs(mean), na.rm = TRUE) %>% 
  rename(Total_fished_fmi_wadj = Total_fished,
         disp_fished_fmi_wadj = disp_fished,
         coi_disp_fished_fmi_wadj = coi_disp_fished,
         unequal_fish_fmi_wadj = unequal_fish,
         unequal_fish_hs_fmi_wadj = unequal_fish_hs,
         unequal_fishing_in_waters_fmi_wadj = unequal_fishing_in_waters,
         rev_fish_fmi_wadj = rev_fish,
         rev_fishing_in_waters_fmi_wadj = rev_fishing_in_waters,
         unequal_fish_eez_fmi_wadj = unequal_fish_eez) %>% 
  dplyr::select(-Year) 
```

```{r}
sub_fish_fmi<-fish_fmi %>% 
  filter(Year %in% yoi) %>% 
  dplyr::select(-prop_disp, -prop_unequal_disp, -prop_rev_disp) %>% 
  group_by(Fishing_ISO3) %>% 
  summarise_all(funs(mean), na.rm = TRUE) %>% 
  rename(Total_fished_fmi = Total_fished,
         disp_fished_fmi = disp_fished,
         coi_disp_fished_fmi = coi_disp_fished,
         unequal_fish_fmi = unequal_fish,
         unequal_fish_hs_fmi = unequal_fish_hs,
         unequal_fishing_in_waters_fmi = unequal_fishing_in_waters,
         rev_fish_fmi = rev_fish,
         rev_fishing_in_waters_fmi = rev_fishing_in_waters,
         unequal_fish_eez_fmi = unequal_fish_eez) %>% 
  dplyr::select(-Year) 
```

# Join all together

```{r}
all_dat<-full_join(sub_trade, sub_fish, by = c("Import_ISO3" = "Fishing_ISO3")) %>% 
  full_join(., sub_fish_fmi_wadj, by = c("Import_ISO3" = "Fishing_ISO3")) %>% 
  full_join(., sub_fish_fmi, by = c("Import_ISO3" = "Fishing_ISO3")) %>% 
  rename(Country = Import_ISO3) %>% 
  full_join(., me_dat, by = c("Country" = "Export_ISO3")) %>% 
  full_join(., fmi_dat, by = c("Country" = "Export_ISO3"))
  #filter(!is.na(ME_score) == T) %>% 
  
```

Need to double check trade - way off from current analysis

```{r}
final<-all_dat %>% 
  mutate(Country.Name = countrycode(Country, "iso3c", "country.name"),
         Country.Name = ifelse(Country == "ANT", "Netherlands Antilles",
                               ifelse(Country == "SCG", "Serbia and Montenegro",
                                      ifelse(Country == "HighSeas", "High seas", Country.Name)))) %>% 
  rename(ISO3 = Country) %>% 
  dplyr::select(-c(X, rev_fish, rev_fishing_in_waters, rev_fish_fmi_wadj, rev_fishing_in_waters_fmi_wadj, rev_fish_fmi, rev_fishing_in_waters_fmi)) %>% 
  rename(ME_quant = quantile)
    
  #%>% 
 # dplyr::select(Country.Name, ISO3, ME_score, quantile, imports,  unequal_imports,disp_fished, unequal_fish, unequal_fish_eez, unequal_fish_hs, unequal_fishing_in_waters, disp_fished_fmi_wadj, unequal_fish_fmi_wadj, unequal_fish_eez_fmi_wadj, unequal_fish_hs_fmi_wadj, unequal_fishing_in_waters_fmi_wadj, FMI_weight_adj) %>% 
 # rename(ME_quant = quantile)
```

```{r}
write.csv(final, here("output_data/Supplementary_Table1.csv"))
```
